requirements.txt
asgiref==3.8.1
astroid==3.3.3
autopep8==2.3.1
black==24.8.0
click==8.1.7
dill==0.3.8
Django==4.2.16
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-yasg==1.21.7
flake8==7.1.1
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
